feat: Implement language selection and translation updates
- Added Language enum and SettingsModel for managing user settings.
- Integrated language selection in SettingsTab with segmented buttons.
- Updated TimerProvider to handle language-specific notifications.
- Enhanced translation handling in various UI components.
- Refactored exercise count display to support singular/plural forms based on the number.
- Improved UI consistency and user experience across settings and workout screens.

+++ lib/utils/progress_bar.dart
@@ -3,8 +3,8 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 
 class ProgressBar extends StatelessWidget {
-  final double value;
-  final double maxValue;
+  final int value;
+  final int maxValue;
   const ProgressBar({super.key, required this.value, required this.maxValue});
 
   @override
